moved main class to Separate folder
