+++ .gitignore
@@ -15,3 +15,6 @@ functions/__pycache__/test_func.cpython-310-pytest-7.1.2.pyc
 functions/__pycache__/tools.cpython-310.pyc
 tests/__pycache__/test_func.cpython-310-pytest-7.1.2.pyc
 functions/__pycache__/classes.cpython-310.pyc
+classes/__pycache__/__init__.cpython-310.pyc
+.gitignore
+classes/__pycache__/main_class.cpython-310.pyc
